third_party/blink/web_tests/fast/forms/calendar-picker/week-picker-mouse-operations-expected.txt

Tests if week picker mouse operations work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W01"
PASS highlightedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
Check that hovering over an entry highlights it.
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "1999-W51"
PASS highlightedDayCells() is "1999-12-26"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "1999-W52"
PASS highlightedDayCells() is "1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01,2000-01-02"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W01"
PASS highlightedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W05"
PASS highlightedDayCells() is "2000-01-31,2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05"
Check that mouse click sets the value.
PASS currentMonth() is "2000-01"
PASS selectedValue() is "1999-W52"
PASS selectedDayCells() is "1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01,2000-01-02"
PASS highlightedValue() is "1999-W52"
PASS highlightedDayCells() is "1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01,2000-01-02"
PASS document.getElementById("week").value is "1999-W52"
Check that clicking the month popup button opens the month popup.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS popupWindow.global.picker.monthPopupView.isVisible is true
PASS highlightedMonthButton() is "2000-01"
Check that hovering over a month button highlights it.
PASS highlightedMonthButton() is "2000-10"
Check that month wheel scrolls the list in the month popup.
PASS checkYearListViewScrollOffset() > 0 is true
PASS checkYearListViewScrollOffset() < 0 is true
Check that clicking the month popup sets the month.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS currentMonth() is "2000-06"
Check that clicking the month popup button opens the month popup.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS popupWindow.global.picker.monthPopupView.isVisible is true
PASS highlightedMonthButton() is "2000-06"
Check that hovering over a month button highlights it.
PASS highlightedMonthButton() is "2000-10"
Check that month wheel scrolls the list in the month popup.
PASS checkYearListViewScrollOffset() > 0 is true
PASS checkYearListViewScrollOffset() < 0 is true
Check that clicking an year list cell opens it.
PASS checkYearListViewScrollOffset() > 0 is true
PASS highlightedMonthButton() is "1999-01"
Check that clicking the month popup sets the month.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS currentMonth() is "1999-06"
PASS successfullyParsed is true

TEST COMPLETE
